matchers: move `RepoPathTree` to `repo_path` module
I plan to use this type in at least `TreeBuilder`.

I renamed some `dir` variables to `path` since I don't think they have
to be directories.
